+++ packages/feature-quest-hub/src/ui/components/LeGameLevel.styles.js
@@ -1,0 +1,11 @@
+import { css } from "lit";
+
+export const gameLevelStyles = css`
+  :host {
+    display: block;
+    position: relative;
+    width: 100vw;
+    height: 100vh;
+    background-color: #000;
+  }
+`;